+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file, per [the Ke
 ## [Unreleased]
 
 ### Added
-* Added compatibility with debug-bar-wpprobe plugin
+* Added compatibility with debug-bar-elasticprobe plugin
 
 ### Changed
 * Changed default ES primary shard count to 3
+* Changed plugin name to ElasticProbe and refactored accordingly
 
 ### Fixed
 * Status report page